fix(fork): explicit rlim_t to int cast for resource limit comparison
- Add explicit (int) cast for parent->rlimits[RLIMIT_NPROC].rlim_cur
- Fixes sign comparison warning between int and rlim_t (unsigned long long)
- Same fix applied consistently in both iox_fork() and iox_vfork()

Fulfills: VAL-EPIC2-001, VAL-EPIC2-003

+++ ixland-system/kernel/task/fork.c
@@ -93,7 +93,7 @@ pid_t iox_fork(void) {
         child_count++;
         c = c->next_sibling;
     }
-    if (child_count >= parent->rlimits[RLIMIT_NPROC].rlim_cur) {
+    if (child_count >= (int)parent->rlimits[RLIMIT_NPROC].rlim_cur) {
         pthread_mutex_unlock(&parent->lock);
         errno = EAGAIN;
         return -1;
